Prepare to develop command line handling code
